refactor(15-02): split config.rs into config/{mod,types,overrides,validate}.rs
- types.rs hosts data shapes (Config, DirectoryName, DirectoryConfig,
  DirectoryType, DirectoryRole, GitRef, BackupConfig)
- validate.rs hosts Config::validate (Cases A/B/C overlap detection)
- overrides.rs hosts apply_machine_overrides + warn_unknown_overrides +
  format_override_validation_error (PORT-01..05)
- mod.rs hosts Config::load/load_or_default/save/save_checked/
  load_with_overrides + tome_home/XDG helpers (S3 lock for save_checked)
- Tilde helpers (expand_tilde) lifted into paths.rs per CONTEXT.md
  Claude's Discretion; mod.rs re-exports it so existing
  crate::config::expand_tilde call sites continue to compile.

HARD-03 (#487). 845 tests preserved.
